+++ nvtVoronoi.cpp
@@ -1,0 +1,117 @@
+#include "std_include.h"
+
+#include "cuda_runtime.h"
+#include "cuda_profiler_api.h"
+
+#define ENABLE_CUDA
+
+#include "Simulation.h"
+#include "voronoiQuadraticEnergy.h"
+#include "NoseHooverChainNVT.h"
+
+/*!
+This file explores integrating the Nose-Hoover equations of motion
+*/
+int main(int argc, char*argv[])
+{
+    //...some default parameters
+    int numpts = 200; //number of cells
+    int USE_GPU = 0; //0 or greater uses a gpu, any negative number runs on the cpu
+    int c;
+    int tSteps = 5; //number of time steps to run after initialization
+    int initSteps = 1; //number of initialization steps
+
+    Dscalar dt = 0.01; //the time step size
+    Dscalar p0 = 3.8;  //the preferred perimeter
+    Dscalar a0 = 1.0;  // the preferred area
+    Dscalar v0 = 0.1;  // the self-propulsion
+
+    //The defaults can be overridden from the command line
+    while((c=getopt(argc,argv,"n:g:m:s:r:a:i:v:b:x:y:z:p:t:e:")) != -1)
+        switch(c)
+        {
+            case 'n': numpts = atoi(optarg); break;
+            case 't': tSteps = atoi(optarg); break;
+            case 'g': USE_GPU = atoi(optarg); break;
+            case 'i': initSteps = atoi(optarg); break;
+            case 'e': dt = atof(optarg); break;
+            case 'p': p0 = atof(optarg); break;
+            case 'a': a0 = atof(optarg); break;
+            case 'v': v0 = atof(optarg); break;
+            case '?':
+                    if(optopt=='c')
+                        std::cerr<<"Option -" << optopt << "requires an argument.\n";
+                    else if(isprint(optopt))
+                        std::cerr<<"Unknown option '-" << optopt << "'.\n";
+                    else
+                        std::cerr << "Unknown option character.\n";
+                    return 1;
+            default:
+                       abort();
+        };
+
+    clock_t t1,t2; //clocks for timing information
+    bool reproducible = true; // if you want random numbers with a more random seed each run, set this to false
+    //check to see if we should run on a GPU
+    bool initializeGPU = true;
+    if (USE_GPU >= 0)
+        {
+        bool gpu = chooseGPU(USE_GPU);
+        if (!gpu) return 0;
+        cudaSetDevice(USE_GPU);
+        }
+    else
+        initializeGPU = false;
+
+    //define the equation of motion to be used
+    shared_ptr<NoseHooverChainNVT> nvt = make_shared<NoseHooverChainNVT>(numpts,4);
+    nvt->setT(v0);
+    //define a voronoi configuration with a quadratic energy functional
+    shared_ptr<VoronoiQuadraticEnergy> vm  = make_shared<VoronoiQuadraticEnergy>(numpts,1.0,p0,reproducible);
+
+    //combine the equation of motion and the cell configuration in a "Simulation"
+    SimulationPtr sim = make_shared<Simulation>();
+    sim->setConfiguration(vm);
+    sim->addUpdater(nvt,vm);
+    //set the time step size
+    sim->setIntegrationTimestep(dt);
+    //initialize Hilbert-curve sorting... can be turned off by commenting out this line or seting the argument to a negative number
+    sim->setSortPeriod(initSteps/10);
+    //set appropriate CPU and GPU flags
+    sim->setCPUOperation(!initializeGPU);
+    sim->setReproducible(reproducible);
+
+    //run for a few initialization timesteps
+    printf("starting initialization\n");
+    for(int ii = 0; ii < initSteps; ++ii)
+        {
+        sim->performTimestep();
+        };
+
+    printf("Finished with initialization\n");
+    cout << "current q = " << vm->reportq() << endl;
+    //the reporting of the force should yield a number that is numerically close to zero.
+    vm->reportMeanCellForce(false);
+
+    //run for additional timesteps, and record timing information
+    t1=clock();
+    Dscalar meanT = 0.0;
+    for(int ii = 0; ii < tSteps; ++ii)
+        {
+        meanT += nvt->kineticEnergy/(2.*numpts);
+        if(ii%100 ==0)
+            {
+            printf("timestep %i\t\t energy %f \t T %f \n",ii,vm->computeEnergy(),nvt->kineticEnergy/(2.*numpts));
+            };
+        sim->performTimestep();
+        };
+    t2=clock();
+    Dscalar steptime = (t2-t1)/(Dscalar)CLOCKS_PER_SEC/tSteps;
+    cout << "timestep ~ " << steptime << " per frame; " << endl;
+    cout << vm->reportq() << endl;
+    cout << "<T> = " << meanT / tSteps << endl;
+
+    if(initializeGPU)
+        cudaDeviceReset();
+    return 0;
+};